move getHost to gethost.cpp
