+++ src/gethost.hpp
@@ -1,1 +1,3 @@
+typedef std::pair<std::string, unsigned int> HOST;
 std::string GetLANIPAddress();
+HOST getHost(std::string filename);
