+++ src/main.cpp
@@ -12,7 +12,6 @@
 #include "window.hpp"
 #include "gethost.hpp"
 
-
 #pragma comment(lib, "ws2_32.lib")
 #pragma comment(lib, "iphlpapi.lib")
 
@@ -21,38 +20,12 @@ char QR_0_STR[2] = {(char)32, (char)32};
 
 #define QR_FALSE
 
-typedef std::pair<std::string, unsigned int> HOST;
-
 std::string getParentDirectory(const std::string& fullPath) {
     size_t lastSlashPos = fullPath.find_last_of("\\/");
     if (lastSlashPos != std::string::npos) {
         return fullPath.substr(0, lastSlashPos + 1); // Include trailing slash
     }
     return "";
-}
-
-HOST getHost(std::string filename)
-{
-    HOST host;
-    
-    std::ifstream hostFile(filename);
-    if (!hostFile.is_open())
-    {
-        std::cout << "Unable to open host.txt\nAttempting to get primary adapter ip...";
-        host.first = GetLANIPAddress();
-        if (host.first.empty())
-        {
-            std::cout << "FAILED\n";
-            std::cerr << "Unable to open host.txt, and failed to determine primary adapter ip. Ensure that host.txt exists in the same directory as the Axon executable.";
-        }
-    }
-
-    hostFile >> host.first;
-    std::string portbuf;
-    hostFile >> portbuf;
-    host.second = std::stoi(portbuf);
-
-    return host;
 }
 
 int main(int argc, char* argv[]) {
